# swagger.yaml
---
swagger: '2.0'
info:
  description: |
    This is a sample Petstore server.  You can find 
    out more about Swagger at 
    [http://swagger.io](http://swagger.io) or on 
    [irc.freenode.net, #swagger](http://swagger.io/irc/).
  version: 1.0.0
  title: Swagger Bookstore
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# host: petstore.swagger.io
# basePath: /v2
tags:
  - name: author
    description: Everything about your Authors
    externalDocs:
      description: Find out more
      url: http://swagger.io

# schemes:
# - http
paths:
  /authors:
    post:
      tags:
        - author
      summary: Add a new author to the store
      operationId: addAuthor
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Author object that needs to be added to the store
          required: true
          schema:
            $ref: '#/definitions/Author'
      responses:
        404:
          description: Author not found
    get:
      tags:
        - author
      summary: Get all authors in repository
      description: Returns all authors
      operationId: getAllAuthors
      produces:
        - application/json
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/AuthorList'
        404:
          description: Author not found

  /authors/{authorId}:
    put:
      tags:
        - author
      summary: Update an existing author
      operationId: updateAuthor
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: authorId
          description: Author object that needs to be updated
          required: true
          type: string
        - in: body
          name: author
          description: Author object that needs to be updated
          required: true
          schema:
            $ref: '#/definitions/Author'
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/Author'
        404:
          description: Author not found
    get:
      tags:
        - author
      summary: Find author by ID
      description: Returns a single author
      operationId: getAuthorById
      produces:
        - application/json
      parameters:
        - name: authorId
          in: path
          description: ID of author to return
          required: true
          type: string
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/Author'
        404:
          description: Author not found
    delete:
      tags:
        - author
      summary: Deletes a author
      operationId: deleteAuthor
      produces:
        - application/json
      parameters:
        - name: authorId
          in: path
          description: Author id to delete
          required: true
          type: string
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/Author'
        404:
          description: Author not found


definitions:

  AuthorList:
    type: array
    items:
      $ref: '#/definitions/Author'
    example:
      - id: string
        name: string
        country: string
        birthDate: string
      - id: string
        name: string
        country: string
        birthDate: string

  Author:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      country:
        type: string
      birthDate:
        type: string

externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io